refactor: rename locales folder
