+++ src/client/locales/i18n.types.ts
@@ -1,5 +1,5 @@
 import 'react-i18next';
-import en from './locales/en.json';
+import en from './en.json';
 
 declare module 'react-i18next' {
   interface CustomTypeOptions {
